frontend/gamma/tests/tests/Bookmarklet/Fail.html

<!--

Copyright 2008-2015 Clipperz Srl

This file is part of Clipperz, the online password manager.
For further information about its features and functionalities please
refer to http://www.clipperz.com.

* Clipperz is free software: you can redistribute it and/or modify it 
  under the terms of the GNU Affero General Public License as published
  by the Free Software Foundation, either version 3 of the License, or 
  (at your option) any later version.

* Clipperz is distributed in the hope that it will be useful, but 
  WITHOUT ANY WARRANTY; without even the implied warranty of 
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  See the GNU Affero General Public License for more details.

* You should have received a copy of the GNU Affero General Public
  License along with Clipperz. If not, see http://www.gnu.org/licenses/.

-->

<html>
<head>
	<title>Bookmarlet TEST - fail</title>
	<script type="text/javascript" src="../../../js/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../../SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="../../SimpleTest/test.css">

    <!-- link rel="stylesheet" type="text/css" href="./Bookmarklet.css" -->
	<script type='text/javascript' src='../../../js/Bookmarklet.js'></script>
	<script type='text/javascript' src='../../../../_build/_scratch/bookmarklet_test.js'></script>
</head>
<body>

<!-- TWITTER Login Form -->
<!-- form id="signin" action="https://twitter.com/sessions" method="post">
	<input id="authenticity_token" type="hidden" value="a3a9abfa089122e4b437736313e54f2a25911101" name="authenticity_token"/>
	<fieldset class="common-form standard-form">
		<legend>Please sign in</legend>
		<p>
			<label class="inside" for="username" tabindex="1">user name or email address:</label>
			<input id="username" type="text" title="username" value="" name="session[username_or_email]"/>
		</p>
		<p>
			<label class="inside" for="password" tabindex="2">password:</label>
			<input id="password" type="password" title="password" value="" name="session[password]"/>
		</p>
		<p class="remember"></p>
		<p class="submit"></p>
		<p class="forgot"></p>
		<p class="complete"></p>
	</fieldset>
</form -->

<!-- ###################################################################### -->
<!--
<div id="clipperzBookmarklet" style="">
	<div id="clipperzBookmarkletResult" style=""><p>The direct login configuration has been collected.</p></div>
	<div id="clipperzBookmarletButton" style=""></div>
	<div id="clipperzBookmarletAfterCopyHint" style=""><p>Now you can return to the Clipperz main interface and create a new card</p></div>
</div>
-->
<!-- 
<div id="clipperzBookmarklet">
	<div id="clipperzBookmarkletResult"><div id="clipperzBookmarkletResultIcon"></div><p id="clipperzBookmarkletResultText"></p></div>
	<div id="clipperzBookmarletButton"></div>
	<div id="clipperzBookmarletAfterCopyHint"><p id="clipperzBookmarkletHintText"></p></div>
</div>
-->


<!-- ###################################################################### -->

<pre id="test">
<script type="text/javascript">
try {
	runBookmarklet();
//	var	parameters;
//	runBookmarklet();

//	parameters = getLoginFormConfiguration();
	
//	SimpleTest.ok(parameters != null,		"The bookmarklet returns something useful");
//	SimpleTest.ok(bookmarkletClip != null,	"The clipboard bridge has been created");
//	SimpleTest.ok(bookmarkletClip.ready,	"The clipboard bridge has been created AND is ready");
	SimpleTest.ok(true, "no exception raised");

} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
}

</script>
</pre>
</body>
</html>
